requirements.txt
PyJWT==2.10.1
tabulate==0.9.0